requirements.txt
Pillow==9.5.0
pyproj==3.5.0
python-telegram-bot==20.3
Requests==2.31.0
selenium==4.9.1
